requirements.txt
astroid==2.2.5
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
pylint==2.3.1
python-dateutil==2.8.0
six==1.12.0
typed-ast==1.3.4
wrapt==1.11.1
